moved progressmodel out of class
