+++ app/src/main/java/com/vanced/manager/ui/dialogs/AppDownloadDialog.kt
@@ -17,8 +17,7 @@ import com.vanced.manager.core.downloader.MusicDownloader.downloadMusic
 import com.vanced.manager.core.downloader.VancedDownloader.downloadVanced
 import com.vanced.manager.core.ui.base.BindingDialogFragment
 import com.vanced.manager.databinding.DialogAppDownloadBinding
-import com.vanced.manager.utils.DownloadHelper.downloadProgress
-import com.vanced.manager.utils.applyAccent
+import com.vanced.manager.utils.*
 
 class AppDownloadDialog : BindingDialogFragment<DialogAppDownloadBinding>() {
 
@@ -82,27 +81,25 @@ class AppDownloadDialog : BindingDialogFragment<DialogAppDownloadBinding>() {
     }
 
     private fun DialogAppDownloadBinding.bindDownloadProgress() {
-        with(downloadProgress) {
-            downloadProgress.observe(viewLifecycleOwner) {
-                appDownloadProgressbar.progress = it
-            }
-            installing.observe(viewLifecycleOwner) { installing ->
-                appDownloadProgressbar.isVisible = !installing
-                appInstallProgressbar.isVisible = installing
-                appDownloadFile.isVisible = !installing
-                appDownloadCancel.isEnabled = !installing
-                appDownloadCancel.setOnClickListener {
-                    if (installing) {
-                        return@setOnClickListener
-                    }
-                    currentDownload?.cancel()
-                    downloadProgress.value = 0
-                    dismiss()
+        downloadProgress.observe(viewLifecycleOwner) {
+            appDownloadProgressbar.progress = it
+        }
+        installing.observe(viewLifecycleOwner) { installing ->
+            appDownloadProgressbar.isVisible = !installing
+            appInstallProgressbar.isVisible = installing
+            appDownloadFile.isVisible = !installing
+            appDownloadCancel.isEnabled = !installing
+            appDownloadCancel.setOnClickListener {
+                if (installing) {
+                    return@setOnClickListener
                 }
+                currentDownload?.cancel()
+                downloadProgress.value = 0
+                dismiss()
             }
-            downloadingFile.observe(viewLifecycleOwner) {
-                appDownloadFile.text = it
-            }
+        }
+        downloadingFile.observe(viewLifecycleOwner) {
+            appDownloadFile.text = it
         }
     }
